Extract Patchlevel correctly from the Appveyor tag into Vims version.h
This uses a different approach to get the correct patchlevel. Instead of
extracting the patch from version.c, it uses the tag name for that. This
matters for Patches like 9.0.0049, which previously used the patchlevel
of 49 but it should be '0049'

This patchlevel will be used as DisplayVersion and stored in the Windows
registry. winget makes use of the DisplayVersion field and compares that
against the version in its repository (which will again use the tag
version).

So this makes sure we use the proper version string for the installed
Windows Vim installer (also visible in the control panel of add/remove
programs).

closes #276